delete: removido metodos de adicionar produtos
